pyproject.toml
[project]
name = "hypothesis-mcp"
version = "0.1.0"
description = "MCP server for the Hypothesis annotation service — gives Claude access to your annotations and PDFs"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Ankit Goyal" },
]
keywords = ["mcp", "hypothesis", "annotations", "pdf", "claude", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.1",
    "pdfplumber>=0.11.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/imankgoyal/hypothesis-mcp"
Repository = "https://github.com/imankgoyal/hypothesis-mcp"
Issues = "https://github.com/imankgoyal/hypothesis-mcp/issues"

[project.scripts]
hypothesis-mcp = "hypothesis_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hypothesis_mcp"]
